Make github-watcher fetch more pages of repos.
We do this parsing the Link header, and keep fetching so long as we have a "next" link. We emit an end event at the end as well (mainly used in the test).

The test, for now, is live, which means it is a bit slow, hence the increased timeout for `npm test`